pyproject.toml
[tool.poetry]
name = "homework-30"
version = "0.1.0"
description = ""
authors = ["spbproger <zemli2016@yandex.ru>"]
readme = "README.md"
packages = [{include = "homework_30"}]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
